+++ Kambaz/Courses/routes.js
@@ -20,14 +20,36 @@ export default function CourseRoutes(app, db) {
       res.sendStatus(401);
       return;
     }
+    if (currentUser.role === "STUDENT") {
+      res.sendStatus(403);
+      return;
+    }
     const newCourse = dao.createCourse(req.body, currentUser._id);
     res.json(newCourse);
   };
   const deleteCourse = (req, res) => {
+    const currentUser = req.session.currentUser;
+    if (!currentUser) {
+      res.sendStatus(401);
+      return;
+    }
+    if (currentUser.role === "STUDENT") {
+      res.sendStatus(403);
+      return;
+    }
     dao.deleteCourse(req.params.courseId);
     res.sendStatus(200);
   };
   const updateCourse = (req, res) => {
+    const currentUser = req.session.currentUser;
+    if (!currentUser) {
+      res.sendStatus(401);
+      return;
+    }
+    if (currentUser.role === "STUDENT") {
+      res.sendStatus(403);
+      return;
+    }
     const updated = dao.updateCourse(req.params.courseId, req.body);
     res.json(updated);
   };
